Fix warning

+++ Shared/Include/Solver.h
@@ -13,7 +13,7 @@ namespace Sandbox
 		virtual ~INodeList() {}
 
 		virtual void AddNode(ISolverNode* pNode) = 0;
-		virtual int Count() const = 0;
+		virtual size_t Count() const = 0;
 		virtual const ISolverNode* At(int idx) const = 0;
 	};
